plugins/physioqm/app/main.py
"""
PhysioQM Plugin - gRPC Server
==============================

Quantum-informed PBPK prediction service

Author: Agourakis
Created: 2025-10-31T01:30:00Z
Purpose: Real science - mechanistic interpretability
Darwin Indexed: Auto-indexed via MCP
"""

import grpc
from grpc import aio
import asyncio
import logging
from datetime import datetime, timezone
import sys
from pathlib import Path

# Add proto path
sys.path.insert(0, str(Path(__file__).parent.parent))

try:
    from protos.physioqm.v1 import physioqm_pb2, physioqm_pb2_grpc
    PROTOS_AVAILABLE = True
except ImportError:
    logging.warning("Proto files not found - run generate_protos.sh")
    PROTOS_AVAILABLE = False

logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger("physioqm.server")


class PhysioQMServicer(physioqm_pb2_grpc.PhysioQMServiceServicer):
    """
    PhysioQM gRPC service implementation
    
    Implements:
    - PredictPBPK: Main prediction endpoint
    - ExtractQuantumFeatures: DFT calculations
    - DockToCYP450: Molecular docking
    - PredictTransporters: Transporter predictions
    - GenerateVisualMaps: ESP/Fukui maps
    - ExplainPrediction: Interpretability
    """
    
    def __init__(self):
        logger.info("🧬 Initializing PhysioQM Service...")
        
        # Load models (placeholder - will implement)
        self.model = None
        self.quantum_service = None
        self.enzyme_service = None
        self.transporter_service = None
        self.visual_service = None
        
        logger.info("✅ PhysioQM Service initialized")
    
    async def PredictPBPK(self, request, context):
        """
        Main prediction endpoint
        
        Args:
            request: PredictRequest with SMILES
            context: gRPC context
            
        Returns:
            PBPKPredictionResponse with Fu, Vd, Clearance
        """
        logger.info(f"📊 PredictPBPK called: {request.smiles[:50]}...")
        
        try:
            # TODO: Implement full pipeline
            # 1. Extract quantum features
            # 2. Dock to CYP450
            # 3. Predict transporters
            # 4. Generate visual maps (if requested)
            # 5. Model inference
            # 6. Interpret results (if requested)
            
            # Placeholder response
            response = physioqm_pb2.PBPKPredictionResponse(
                success=False,
                error="Not implemented yet - under development for REAL SCIENCE",
                metadata={
                    "timestamp": datetime.now(timezone.utc).isoformat(),
                    "version": "0.1.0-pilot"
                }
            )
            
            return response
        
        except Exception as e:
            logger.error(f"PredictPBPK failed: {e}")
            return physioqm_pb2.PBPKPredictionResponse(
                success=False,
                error=str(e)
            )
    
    async def ExtractQuantumFeatures(self, request, context):
        """Extract quantum features via DFT"""
        logger.info(f"⚛️  ExtractQuantumFeatures: {request.smiles[:50]}...")
        
        # TODO: Implement DFT calculation
        return physioqm_pb2.QuantumFeaturesResponse(
            success=False,
            error="Not implemented yet"
        )
    
    async def DockToCYP450(self, request, context):
        """Dock molecule to CYP450 isoforms"""
        logger.info(f"🔬 DockToCYP450: {request.smiles[:50]}...")
        
        # TODO: Implement AutoDock-GPU
        return physioqm_pb2.DockingScoresResponse(
            success=False,
            error="Not implemented yet"
        )
    
    async def PredictTransporters(self, request, context):
        """Predict transporter interactions"""
        logger.info(f"🚪 PredictTransporters: {request.smiles[:50]}...")
        
        # TODO: Implement TDC models
        return physioqm_pb2.TransporterResponse(
            success=False,
            error="Not implemented yet"
        )
    
    async def GenerateVisualMaps(self, request, context):
        """Generate ESP and Fukui visual maps"""
        logger.info(f"🎨 GenerateVisualMaps: {request.smiles[:50]}...")
        
        # TODO: Implement ESP/Fukui rendering
        return physioqm_pb2.VisualMapsResponse(
            success=False,
            error="Not implemented yet"
        )
    
    async def ExplainPrediction(self, request, context):
        """Explain prediction mechanistically"""
        logger.info(f"💡 ExplainPrediction: {request.smiles[:50]}...")
        
        # TODO: Implement SHAP + attention analysis
        return physioqm_pb2.ExplanationResponse(
            success=False,
            error="Not implemented yet"
        )
    
    async def GetAttentionWeights(self, request, context):
        """Get attention weights visualization"""
        logger.info(f"👁️  GetAttentionWeights: {request.smiles[:50]}...")
        
        # TODO: Implement attention extraction
        return physioqm_pb2.AttentionResponse(
            success=False,
            error="Not implemented yet"
        )
    
    async def StreamPredictions(self, request_iterator, context):
        """Batch predictions via streaming"""
        logger.info("📡 StreamPredictions called")
        
        # TODO: Implement streaming
        async for request in request_iterator:
            yield physioqm_pb2.PBPKPredictionResponse(
                success=False,
                error="Streaming not implemented yet"
            )
    
    async def HealthCheck(self, request, context):
        """Health check endpoint"""
        logger.info("💚 HealthCheck called")
        
        # Check components
        component_status = {
            "quantum_service": "not_initialized",
            "enzyme_service": "not_initialized",
            "transporter_service": "not_initialized",
            "visual_service": "not_initialized",
            "model": "not_loaded"
        }
        
        all_healthy = False  # Will be True when all components loaded
        
        response = physioqm_pb2.HealthResponse(
            healthy=all_healthy,
            status="degraded" if not all_healthy else "healthy",
            component_status=component_status,
            metrics={
                "uptime_seconds": "0",  # TODO: Track uptime
                "predictions_served": "0",
                "cache_hit_rate": "0.0"
            }
        )
        
        return response


async def serve():
    """Start gRPC server"""
    if not PROTOS_AVAILABLE:
        logger.error("❌ Proto files not available - cannot start server")
        return
    
    port = 50052  # PhysioQM uses 50052 (Darwin Core uses 50051)
    
    server = aio.Server()
    physioqm_pb2_grpc.add_PhysioQMServiceServicer_to_server(
        PhysioQMServicer(), server
    )
    
    server.add_insecure_port(f'[::]:{port}')
    
    logger.info("="*80)
    logger.info("🚀 PhysioQM Plugin - gRPC Server Starting")
    logger.info("="*80)
    logger.info(f"   Port: {port}")
    logger.info(f"   Proto: darwin.physioqm.v1")
    logger.info(f"   Mission: REAL SCIENCE for humanity")
    logger.info("="*80)
    
    await server.start()
    
    logger.info("✅ PhysioQM server ready!")
    
    await server.wait_for_termination()


if __name__ == "__main__":
    asyncio.run(serve())

