plugins/physioqm/requirements.txt
# PhysioQM Plugin Dependencies
# Created: 2025-10-31T01:30:00Z
# Purpose: Real science PBPK prediction

# Core
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2

# ML/DL
torch==2.1.0
torch-geometric==2.4.0
torchvision==0.16.0
transformers==4.35.0

# Quantum Chemistry
pyscf==2.3.0

# Molecular
rdkit==2023.9.1
openbabel==3.1.1

# Docking (will install AutoDock-GPU separately)
# meeko  # For PDBQT preparation

# Cheminformatics
tdc==0.4.1  # Therapeutics Data Commons

# Data
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
PIL==10.1.0

# Utils
tqdm==4.66.1
pyyaml==6.0.1

